Removed the closing </title> tag from the generator to allow for text to be appended to the page title.
The closing </title> tag should be put right above the </head> tag in _includes/body.inc.html, along with any text that should
immediately follow the generated title. This means that the <title> tag will always be the last tag in <head>, however, so anything
else that should go in the <head> tag should be defined above the opening <title> tag in _includes/header.inc.html.